+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 Hi! This is where we document all notable changes, including bug fixes, enhancements, and dependency updates.
 Dates should be in`YYYY-MM-DD` format and versions are in [semantic versioning](http://semver.org/) format.
 
+## 0.1.6 2025-02-09
+
+### Maintenance
+
+- Updated all minor versions of vite, Typescript, Svelte, Supabase.
+
 ## 0.1.5 2025-02-02
 
 ### Added